+++ backend/v1/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class V1Config(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "v1"
